refactor(operator_commands_dashboard): extract memory handlers + agent graph helpers
Move memory_search, memory_graph, classify_agent_layer, build_agent_graph
from routes.rs into a new sibling memory.rs module.

- routes.rs: 5123 -> 4827
- memory.rs (NEW, 306)

Refs #1266
