+++ src/operator_commands_dashboard/mod.rs
@@ -1,6 +1,7 @@
 mod auth;
 mod monitoring;
 mod goals;
+mod memory;
 pub(crate) mod routes;
 
 #[cfg(test)]
