+++ src/operator_commands_dashboard/tests_routes_b.rs
@@ -1,6 +1,7 @@
 #[cfg(test)]
 mod tests_b {
     use crate::operator_commands_dashboard::routes::*;
+    use crate::operator_commands_dashboard::memory::{build_agent_graph, classify_agent_layer};
     use crate::operator_commands_dashboard::routes::TmuxSession;
     use serde_json::json;
     #[test]
